+++ projects/configgen/tests/generators/libretro/crt/LibretroConfigCRT_test.py
@@ -17,8 +17,8 @@ ARCADE_TXT = "/recalbox/share/system/configs/crt/arcade_games.txt"
 
 
 def configureForCrt(emulator: Emulator, crtvideostandard="auto", crtresolutiontype="progressive", crtscreentype="15kHz",
-                    crtadaptor="recalboxrgbdual", crtregion="auto", crtscanlines="", rotation=0, verticalgame=False):
-    emulator.configure(keyValueSettings(""),
+                    crtadaptor="recalboxrgbdual", crtregion="auto", crtscanlines="", rotation=0, verticalgame=False, integerscale=False):
+    emulator.configure(keyValueSettings("").setBool("global.integerscale", integerscale),
                        ExtraArguments("", "", "", "", "", "", "", "", crtvideostandard, crtresolutiontype,
                                       crtscreentype,
                                       crtadaptor, crtregion,
@@ -1409,3 +1409,67 @@ def test_given_mk_then_do_not_use_fullscale_on_rgbjamma(mocker):
     assert libretro_config["aspect_ratio_index"] == '23'
     assert libretro_config["crt_switch_timings_ntsc"] == '"1920 1 80 184 312 254 1 7 3 22 0 0 0 54 0 39052806 1"'
     assert libretro_config["custom_viewport_height_ntsc"] == 256
+
+
+def test_given_mk_then_do_not_use_fullscale_on_rgbjamma(mocker):
+    givenThoseFiles(mocker, {
+        ARCADE_TXT: "mk,fbneo,arcade:254@54.706840,0,256,0",
+        MODES_TXT: "arcade:254@54.706840,1920 1 80 184 312 254 1 7 3 22 0 0 0 54 0 39052806 1,54.706840\ndefault:ntsc:240@60,1920 1 80 184 312 240 1 1 3 16 0 0 0 60 0 38937600 1,60"
+    })
+    emulator = configureForCrt(
+        Emulator(name='fbneo', videoMode='1920x1080', ratio='auto', emulator='libretro', core='fbneo'),
+        crtresolutiontype="progressive", crtvideostandard="ntsc", crtadaptor="recalboxrgbjamma",
+        crtscreentype="15kHz")
+
+    libretro_config = LibretroConfigCRT(CRTConfigParser(), CRTModeOffsetter()).createConfigFor(emulator,
+                                                                                               "/recalbox/share/roms/fbneo/mkyturbo.zip")
+
+    assert libretro_config["aspect_ratio_index"] == '23'
+    assert libretro_config["crt_switch_timings_ntsc"] == '"1920 1 80 184 312 254 1 7 3 22 0 0 0 54 0 39052806 1"'
+    assert libretro_config["custom_viewport_height_ntsc"] == 256
+
+
+def test_given_rgbdual_and_integer_scale_then_integer_scale_horizontal(mocker):
+    givenThoseFiles(mocker, {
+        SYSTEMS_TXT: "snes,snes9x,pal,15kHz,progressive,snes:pal:240@50p,0,0\nsnes,snes9x,ntsc,15kHz,progressive,snes:ntsc:224@60p,0,0",
+        MODES_TXT: "snes:ntsc:224@60p,1920 1 78 192 210 224 1 3 3 16 0 0 0 60 0 37730000 1,60.1\nsnes:pal:240@50p,1920 1 78 192 210 240 1 3 3 16 0 0 0 50 0 37730000 1,50.1"})
+
+    emulator = configureForCrt(
+        Emulator(name='snes', videoMode='1920x1080', ratio='auto', emulator='libretro', core='snes9x'),
+        crtresolutiontype="progressive", crtvideostandard="auto", crtadaptor="recalboxrgbdual",
+        crtscreentype="15kHz", integerscale=True)
+
+    libretro_config = LibretroConfigCRT(CRTConfigParser(), CRTModeOffsetter()).createConfigFor(emulator,"smk.zip")
+
+    assert libretro_config["video_scale_integer"] == '"true"'
+    assert libretro_config["aspect_ratio_index"] == '24'
+
+def test_given_rgbdual_and_no_integer_scale_then_no_integer_scale_horizontal(mocker):
+    givenThoseFiles(mocker, {
+        SYSTEMS_TXT: "snes,snes9x,pal,15kHz,progressive,snes:pal:240@50p,0,0\nsnes,snes9x,ntsc,15kHz,progressive,snes:ntsc:224@60p,0,0",
+        MODES_TXT: "snes:ntsc:224@60p,1920 1 78 192 210 224 1 3 3 16 0 0 0 60 0 37730000 1,60.1\nsnes:pal:240@50p,1920 1 78 192 210 240 1 3 3 16 0 0 0 50 0 37730000 1,50.1"})
+
+    emulator = configureForCrt(
+        Emulator(name='snes', videoMode='1920x1080', ratio='auto', emulator='libretro', core='snes9x'),
+        crtresolutiontype="progressive", crtvideostandard="auto", crtadaptor="recalboxrgbdual",
+        crtscreentype="15kHz", integerscale=False)
+
+    libretro_config = LibretroConfigCRT(CRTConfigParser(), CRTModeOffsetter()).createConfigFor(emulator,"smk.zip")
+
+    assert libretro_config["video_scale_integer"] == '"false"'
+    assert libretro_config["aspect_ratio_index"] == '23'
+
+def test_given_rgbjamma_and_no_integer_scale_then_force_integer_scale_horizontal(mocker):
+    givenThoseFiles(mocker, {
+        SYSTEMS_TXT: "snes,snes9x,pal,15kHz,progressive,snes:pal:240@50p,0,0\nsnes,snes9x,ntsc,15kHz,progressive,snes:ntsc:224@60p,0,0",
+        MODES_TXT: "snes:ntsc:224@60p,1920 1 78 192 210 224 1 3 3 16 0 0 0 60 0 37730000 1,60.1\nsnes:pal:240@50p,1920 1 78 192 210 240 1 3 3 16 0 0 0 50 0 37730000 1,50.1"})
+
+    emulator = configureForCrt(
+        Emulator(name='snes', videoMode='1920x1080', ratio='auto', emulator='libretro', core='snes9x'),
+        crtresolutiontype="progressive", crtvideostandard="auto", crtadaptor="recalboxrgbjamma",
+        crtscreentype="15kHz", integerscale=False)
+
+    libretro_config = LibretroConfigCRT(CRTConfigParser(), CRTModeOffsetter()).createConfigFor(emulator,"smk.zip")
+
+    assert libretro_config["video_scale_integer"] == '"true"'
+    assert libretro_config["aspect_ratio_index"] == '24'